requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-crontab==0.7.1
google-auth==2.3.0
google-auth-oauthlib==0.4.6
gspread==4.0.1
httplib2==0.20.1
idna==3.3
numpy==1.21.3
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.3.4
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7